Overhaul url module
