+++ app/current/src/main/scala/io/scalajs/nodejs/url/package.scala
@@ -10,6 +10,7 @@ import scala.scalajs.js.|
   */
 package object url {
 
+  @deprecated("", "")
   type Iterable[A] = js.Array[A] | Iterator[A]
 
 }
